# Ramp up frontend and backend performance with compression and caching