--- a/packages/backend/package.json
+++ b/packages/backend/package.json
@@ -11,9 +11,11 @@
     "@corona/api": "^0.1.0",
     "@corona/pipeline": "^0.1.0",
     "@corona/utils": "^0.1.0",
+    "compression": "^1.7.4",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "lru-cache": "^5.1.1"
+    "lru-cache": "^5.1.1",
+    "winston": "^3.2.1"
   },
   "scripts": {
     "build": "tsc",
@@ -26,10 +28,12 @@
     "lint": "eslint . --ext .ts,.tsx"
   },
   "devDependencies": {
+    "@types/compression": "^1.7.0",
     "@types/cors": "^2.8.6",
     "@types/express": "^4.17.3",
     "@types/lru-cache": "^5.1.0",
     "@types/node": "^12.12.31",
+    "@types/winston": "^2.4.4",
     "nodemon": "^2.0.2",
     "pm2": "^4.2.3",
     "typescript": "^3.8.3"