--- a/packages/backend/src/index.ts
+++ b/packages/backend/src/index.ts
@@ -1,9 +1,12 @@
 import express from "express";
 import { ORIGIN, PORT } from "@corona/api";
+import compression from "compression";
 import { setRoutes } from "./routes";
 import { configureSecurity } from "./security/configureSecurity";
 
 const app = express();
+
+app.use(compression());
 
 configureSecurity(app);
 setRoutes(app);