--- a/packages/frontend/package.json
+++ b/packages/frontend/package.json
@@ -63,6 +63,7 @@
     "@types/reselect": "^2.2.0",
     "@types/topojson-client": "^3.0.0",
     "@types/webpack-env": "^1.15.0",
+    "compression-webpack-plugin": "^3.1.0",
     "css-loader": "^3.4.0",
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^3.2.0",