--- a/packages/frontend/src/index.html
+++ b/packages/frontend/src/index.html
@@ -5,10 +5,6 @@
     <meta charset="UTF-8" />
     <title>Coronavirus dashboard</title>
 
-    <!-- Dependencies -->
-    <script crossorigin src="https://unpkg.com/react@16/umd/react.production.min.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js"></script>
-
     <link href="https://fonts.googleapis.com/css?family=Open+Sans&display=swap" rel="stylesheet">
 
     <style>