--- a/packages/frontend/src/utils/mapDataCache.ts
+++ b/packages/frontend/src/utils/mapDataCache.ts
@@ -3,15 +3,17 @@ import { json } from "d3-fetch";
 
 const CACHE = new LRU({ max: 2 });
 
-export async function getTopology(topologyLocation: string): Promise<any> {
+export async function getTopology(topologyLocation: string, setLoading: (isLoading: boolean) => void): Promise<any> {
     const item = CACHE.get(topologyLocation);
 
     if (item !== undefined) {
         return item;
     }
 
+    setLoading(true);
     const topologyJson = await json(topologyLocation);
     CACHE.set(topologyLocation, topologyJson);
+    setLoading(false);
 
     return topologyJson;
 }