Fix: https://github.com/kadhirvelm/corona/issues/26

Used compression and compression-webpack-plugin to minify everything. It's all performing much, much better. 

At a glance, before compression:
us-county-toplogy.json --> 5.1 mb --> 1.25 seconds
us-state-topology.json --> 904kb --> 267 ms

After compression:
us-county-toplogy.json --> 319 kb --> 567 ms
us-state-topology.json --> 319 kb --> 329 ms

We're also now caching the topology files locally, so it should only call on them as they render and then not call for them again. This both means the initial load time should be lower due to compression of our bundle, but then subsequent renders of the map (both nation and state) should be seamless and in the minimum amount of time possible.